Move Python software into sublist
Given the addition of multiple resources focusing on the large ecosystem
of scientific computing and machine learning libraries in Python make
the summary links a sublist. This is to try and focus the idea that
these are all areas in which to get started with the many tools Python
offers.